requirements.txt
pytest
selenium
assertpy
allure-pytest
requests
pytest-schema
